+++ src/main/java/com/rick/racing/domain/CarName.java
@@ -1,4 +1,4 @@
-package com.rick.racing.model;
+package com.rick.racing.domain;
 
 import java.util.Objects;
