+++ src/main/java/com/rick/racing/domain/TryCount.java
@@ -1,4 +1,4 @@
-package com.rick.racing.model;
+package com.rick.racing.domain;
 
 public class TryCount {
