+++ src/test/java/com/rick/racing/domain/CarNameTest.java
@@ -1,4 +1,4 @@
-package com.rick.racing.model;
+package com.rick.racing.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
